discord_bot/pyproject.toml
[project]
name = "discord_bot"
version = "0.1.0"
requires-python = ">=3.10"

dependencies = [
    "discord.py==2.3.2",
    "python-dotenv==1.0.1",
    "requests==2.32.3"
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio"
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["bot"]

[tool.setuptools.package-dir]
"" = "src"
